requirements.txt
gradio>=4.0.0
numpy>=1.24.0
Pillow>=10.0.0
opencv-python>=4.8.0
requests>=2.31.0
torch>=2.0.0
diffusers>=0.21.0
websocket-client>=1.6.0
ffmpeg-python>=0.2.0
scipy>=1.11.0
